Tool to replace php short tags with long ones
This tool is helpful to convert the files taken from the concrete5
repository, expanding the php short tags into long ones (eg from
`<?` to `<?php` and from `<?=` to `<?php echo `).

This tool works with php tokenizer, so that the replacement is done in
a language-aware way. The only drawback of this is that short tags
inside commented php code are not replaced (but it's shouldn't be a
great problem: it's commented...)